Refactor profilers into their own module.
